fix: store email attachment children as raw blobs
